Assessment entity moved to domain.layer
