+++ Servers/controllers/assessment.ctrl.ts
@@ -27,7 +27,10 @@ import {
   updateQuestionByIdQuery,
   UploadedFile,
 } from "../utils/question.utils";
-import { Assessment, AssessmentModel } from "../models/assessment.model";
+import {
+  Assessment,
+  AssessmentModel,
+} from "../domain.layer/models/assessment/assessment.model";
 import { TopicModel } from "../models/topic.model";
 import { SubtopicModel } from "../models/subtopic.model";
 import { sequelize } from "../database/db";
