+++ Servers/database/db.ts
@@ -1,7 +1,7 @@
 import dotenv from "dotenv";
 import { Sequelize } from "sequelize-typescript";
 import { RoleModel } from "../domain.layer/models/role/role.model";
-import { AssessmentModel } from "../models/assessment.model";
+import { AssessmentModel } from "../domain.layer/models/assessment/assessment.model";
 import { ControlModel } from "../models/control.model";
 import { ControlCategoryModel } from "../models/controlCategory.model";
 import { FileModel } from "../models/file.model";
