+++ Servers/domain.layer/models/assessment/assessment.mock.data.ts
@@ -1,0 +1,11 @@
+import { Assessment } from "./assessment.model";
+
+export const Assessments = (
+  projectId1: number,
+  projectId2: number
+): Assessment[] => {
+  return [
+    { id: 1, project_id: projectId1 },
+    { id: 2, project_id: projectId2 },
+  ];
+};
