--- Servers/mocks/assessment.mock.data.ts
@@ -1,8 +1,0 @@
-import { Assessment } from "../models/assessment.model";
-
-export const Assessments = (projectId1: number, projectId2: number): Assessment[] => {
-  return [
-    { id: 1, project_id: projectId1 },
-    { id: 2, project_id: projectId2 },
-  ]
-};
