+++ Servers/utils/assessment.utils.ts
@@ -1,4 +1,7 @@
-import { Assessment, AssessmentModel } from "../models/assessment.model";
+import {
+  Assessment,
+  AssessmentModel,
+} from "../domain.layer/models/assessment/assessment.model";
 import { sequelize } from "../database/db";
 import { createNewTopicsQuery } from "./topic.utils";
 import { QueryTypes, Transaction } from "sequelize";
@@ -8,7 +11,7 @@ export const getAllAssessmentsQuery = async (): Promise<Assessment[]> => {
     "SELECT * FROM assessments ORDER BY created_at DESC, id ASC",
     {
       mapToModel: true,
-      model: AssessmentModel
+      model: AssessmentModel,
     }
   );
   return assessments;
@@ -22,7 +25,7 @@ export const getAssessmentByIdQuery = async (
     {
       replacements: { id: id },
       mapToModel: true,
-      model: AssessmentModel
+      model: AssessmentModel,
     }
   );
   return result[0];
@@ -36,7 +39,7 @@ export const getAssessmentByProjectIdQuery = async (
     {
       replacements: { project_id: projectId },
       mapToModel: true,
-      model: AssessmentModel
+      model: AssessmentModel,
     }
   );
   return result;
@@ -53,10 +56,14 @@ export const createNewAssessmentQuery = async (
       replacements: { project_id: assessment.project_id },
       mapToModel: true,
       model: AssessmentModel,
-      transaction
+      transaction,
     }
   );
-  const topics = await createNewTopicsQuery(result[0].id!, enable_ai_data_insertion, transaction);
+  const topics = await createNewTopicsQuery(
+    result[0].id!,
+    enable_ai_data_insertion,
+    transaction
+  );
   return { assessment: result[0], topics };
 };
 
@@ -69,12 +76,13 @@ export const updateAssessmentByIdQuery = async (
     `UPDATE assessments SET project_id = :project_id WHERE id = :id RETURNING *`,
     {
       replacements: {
-        project_id: assessment.project_id, id: id
+        project_id: assessment.project_id,
+        id: id,
       },
       mapToModel: true,
       model: AssessmentModel,
       // type: QueryTypes.UPDATE
-      transaction
+      transaction,
     }
   );
   return result.length > 0;
@@ -90,7 +98,7 @@ export const deleteAssessmentByIdQuery = async (
       replacements: { id: id },
       mapToModel: true,
       model: AssessmentModel,
-      transaction
+      transaction,
     }
   );
   return result.length > 0;
