+++ Servers/utils/user.utils.ts
@@ -27,7 +27,7 @@ import { ProjectRiskModel } from "../models/projectRisk.model";
 import { VendorRiskModel } from "../models/vendorRisk.model";
 import { FileModel } from "../models/file.model";
 import { ControlCategoryModel } from "../models/controlCategory.model";
-import { AssessmentModel } from "../models/assessment.model";
+import { AssessmentModel } from "../domain.layer/models/assessment/assessment.model";
 import { TopicModel } from "../models/topic.model";
 import { SubtopicModel } from "../models/subtopic.model";
 import { QuestionModel } from "../models/question.model";
